feat(replication): First iteration on partial sync.

+++ src/server/replica.cc
@@ -290,6 +290,10 @@ std::error_code Replica::HandleCapaDflyResp() {
     return make_error_code(errc::bad_message);
   }
 
+  // If we're syncing a different replication ID, drop the saved LSNs.
+  if (master_context_.master_repl_id != ToSV(LastResponseArgs()[0].GetBuf())) {
+    last_journal_LSNs_.clear();
+  }
   master_context_.master_repl_id = ToSV(LastResponseArgs()[0].GetBuf());
   master_context_.dfly_session_id = ToSV(LastResponseArgs()[1].GetBuf());
   num_df_flows_ = param_num_flows;
@@ -450,25 +454,43 @@ error_code Replica::InitiateDflySync() {
   // Make sure we're in LOADING state.
   CHECK(service_.SwitchState(GlobalState::ACTIVE, GlobalState::LOADING) == GlobalState::LOADING);
 
-  // Flush dbs.
-  JournalExecutor{&service_}.FlushAll();
-
   // Start full sync flows.
   state_mask_.fetch_or(R_SYNCING);
   {
+    // Going out of the way to avoid using std::vector<bool>...
+    auto is_full_sync = std::make_unique<bool[]>(num_df_flows_);
     auto partition = Partition(num_df_flows_);
     auto shard_cb = [&](unsigned index, auto*) {
       for (auto id : partition[index]) {
-        auto ec = shard_flows_[id]->StartFullSyncFlow(sync_block, &cntx_);
+        auto ec = shard_flows_[id]->StartSyncFlow(sync_block, &cntx_, last_journal_LSNs_,
+                                                  is_full_sync.get()[id]);
         if (ec)
           cntx_.ReportError(ec);
       }
     };
-
     // Lock to prevent the error handler from running instantly
     // while the flows are in a mixed state.
     lock_guard lk{flows_op_mu_};
     shard_set->pool()->AwaitFiberOnAll(std::move(shard_cb));
+
+    bool all = true;
+    bool any = false;
+    for (bool is_full : absl::Span<bool>(is_full_sync.get(), num_df_flows_)) {
+      if (is_full)
+        any = true;
+      else
+        all = false;
+    }
+    if (all) {
+      LOG(INFO) << "Will do a full sync!";
+      JournalExecutor{&service_}.FlushAll();
+    } else if (any) {
+      last_journal_LSNs_.clear();
+      cntx_.ReportError(std::make_error_code(errc::state_not_recoverable),
+                        "Won't do a partial sync: some flows must fully resync");
+    } else {
+      LOG(INFO) << "Will do a partial sync!";
+    }
   }
 
   RETURN_ON_ERR(cntx_.GetError());
@@ -600,8 +622,10 @@ error_code Replica::ConsumeDflyStream() {
 }
 
 void Replica::JoinAllFlows() {
+  last_journal_LSNs_.clear();
   for (auto& flow : shard_flows_) {
     flow->JoinFlow();
+    last_journal_LSNs_.push_back(flow->JournalExecutedCount());
   }
 }
 
@@ -625,7 +649,8 @@ error_code Replica::SendNextPhaseRequest(string_view kind) {
   return std::error_code{};
 }
 
-error_code DflyShardReplica::StartFullSyncFlow(BlockingCounter sb, Context* cntx) {
+std::error_code DflyShardReplica::StartSyncFlow(BlockingCounter sb, Context* cntx,
+                                                std::vector<LSN>& lsns, bool& is_full_sync) {
   DCHECK(!master_context_.master_repl_id.empty() && !master_context_.dfly_session_id.empty());
 
   RETURN_ON_ERR(ConnectAndAuth(absl::GetFlag(FLAGS_master_connect_timeout_ms) * 1ms, &cntx_));
@@ -633,8 +658,14 @@ error_code DflyShardReplica::StartFullSyncFlow(BlockingCounter sb, Context* cntx
   VLOG(1) << "Sending on flow " << master_context_.master_repl_id << " "
           << master_context_.dfly_session_id << " " << flow_id_;
 
-  auto cmd = StrCat("DFLY FLOW ", master_context_.master_repl_id, " ",
-                    master_context_.dfly_session_id, " ", flow_id_);
+  std::string cmd;
+  if (lsns.size() > flow_id_ && master_context_.version > DflyVersion::VER1) {
+    cmd = StrCat("DFLY FLOW ", master_context_.master_repl_id, " ", master_context_.dfly_session_id,
+                 " ", flow_id_, " ", lsns[flow_id_]);
+  } else {
+    cmd = StrCat("DFLY FLOW ", master_context_.master_repl_id, " ", master_context_.dfly_session_id,
+                 " ", flow_id_);
+  }
 
   ResetParser(/*server_mode=*/false);
   leftover_buf_.emplace(128);
@@ -648,7 +679,9 @@ error_code DflyShardReplica::StartFullSyncFlow(BlockingCounter sb, Context* cntx
 
   string_view flow_directive = ToSV(LastResponseArgs()[0].GetBuf());
   string eof_token;
-  PC_RETURN_ON_BAD_RESPONSE(flow_directive == "FULL");
+  PC_RETURN_ON_BAD_RESPONSE(flow_directive == "FULL" || flow_directive == "PARTIAL");
+  is_full_sync = flow_directive == "FULL";
+
   eof_token = ToSV(LastResponseArgs()[1].GetBuf());
 
   leftover_buf_->ConsumeInput(read_resp->left_in_buffer);
@@ -680,7 +713,7 @@ error_code DflyShardReplica::StartStableSyncFlow(Context* cntx) {
   return std::error_code{};
 }
 
-void DflyShardReplica::FullSyncDflyFb(const string& eof_token, BlockingCounter bc, Context* cntx) {
+void DflyShardReplica::FullSyncDflyFb(std::string eof_token, BlockingCounter bc, Context* cntx) {
   DCHECK(leftover_buf_);
   io::PrefixSource ps{leftover_buf_->InputBuffer(), Sock()};
 
@@ -722,7 +755,10 @@ void DflyShardReplica::FullSyncDflyFb(const string& eof_token, BlockingCounter b
     leftover_buf_.reset();
   }
 
-  this->journal_rec_executed_.store(loader.journal_offset());
+  if (master_context_.version > DflyVersion::VER0) {
+    CHECK(loader.journal_offset());
+  }
+  this->journal_rec_executed_.store(*loader.journal_offset());
   VLOG(1) << "FullSyncDflyFb finished after reading " << loader.bytes_read() << " bytes";
 }
 
@@ -855,10 +891,11 @@ void DflyShardReplica::ExecuteTxWithNoShardSync(TransactionData&& tx_data, Conte
 }
 
 bool DflyShardReplica::InsertTxToSharedMap(const TransactionData& tx_data) {
-  std::lock_guard lk{multi_shard_exe_->map_mu};
-
+  multi_shard_exe_->map_mu.lock();
   auto [it, was_insert] =
       multi_shard_exe_->tx_sync_execution.emplace(tx_data.txid, tx_data.shard_cnt);
+  multi_shard_exe_->map_mu.unlock();
+
   VLOG(2) << "txid: " << tx_data.txid << " unique_shard_cnt_: " << tx_data.shard_cnt
           << " was_insert: " << was_insert;
   it->second.block.Dec();
